watchtower/wtserver: refactor server handlers into own files
